--- code/examples/pixi-kitchen-sink/jest.config.js
@@ -1,11 +1,0 @@
-const config = require('../../jest.config');
-
-module.exports = {
-  ...config,
-  roots: [__dirname],
-  transform: {
-    ...config.transform,
-    '.*\\.(html)$': '<rootDir>/node_modules/jest-raw-loader',
-  },
-  moduleFileExtensions: [...config.moduleFileExtensions, 'html'],
-};
