--- code/examples/pixi-kitchen-sink/typings.d.ts
@@ -1,6 +1,0 @@
-declare module 'global';
-declare module 'format-json';
-declare module '*.json' {
-  const value: any;
-  export default value;
-}
